Creates directory if it doesn't exist using mkdirp
If the directory doesn't exist when creating a new project, we need to
create it beforehand. The problem stems from the fact that git clone on
windows will, with long directory names, not create the parent directories
needed to create the child. The solution is to create the directory
beforehand with mkdir -p.
Also specified that Python 2.7 is needed on Windows